DataManager refuctorisation
